requirements_cpu.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
alsa-lib=1.2.3=h516909a_0
antlr-python-runtime=4.9.2=py37h89c1867_0
appdirs=1.4.4=pyh9f0ad1d_0
attr=2.4.48=h516909a_0
audioread=2.1.9=py37h89c1867_0
blas=1.0=mkl
brotlipy=0.7.0=py37h5e8e339_1001
bzip2=1.0.8=h7f98852_4
ca-certificates=2021.5.30=ha878542_0
certifi=2021.5.30=py37h89c1867_0
cffi=1.14.6=py37h036bc23_1
chardet=4.0.0=py37h89c1867_1
charset-normalizer=2.0.0=pyhd8ed1ab_0
cpuonly=1.0=0
cryptography=3.4.7=py37h5d9358c_0
cycler=0.10.0=py_2
dbus=1.13.6=h48d8840_2
decorator=5.1.0=pyhd8ed1ab_0
expat=2.4.1=h9c3ff4c_0
ffmpeg=4.3.2=hca11adc_0
fftw=3.3.10=nompi_hcdd671c_100
fontconfig=2.13.1=hba837de_1005
freetype=2.10.4=h0708190_1
gettext=0.19.8.1=h73d1719_1006
glib=2.68.4=h9c3ff4c_1
glib-tools=2.68.4=h9c3ff4c_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
gst-plugins-base=1.18.5=hf529b03_0
gst-plugins-good=1.18.5=h9f60fe5_0
gst-plugins-good-pulseaudio=1.18.4=h3fd9d12_0
gstreamer=1.18.5=h76c114f_0
icu=68.1=h58526e2_0
idna=3.1=pyhd3deb0d_0
intel-openmp=2021.3.0=h06a4308_3350
jack=1.9.18=hfd4fe87_1001
jbig=2.1=h7f98852_2003
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
json-c=0.15=h98cffda_0
kiwisolver=1.3.2=py37h2527ec5_0
krb5=1.19.2=hcc1bbae_0
lame=3.100=h7f98852_1001
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=2.2.1=h9c3ff4c_0
libblas=3.9.0=11_linux64_mkl
libcap=2.48=h7f98852_0
libcblas=3.9.0=11_linux64_mkl
libclang=11.1.0=default_ha53f305_1
libdb=6.2.32=h9c3ff4c_0
libdeflate=1.7=h7f98852_5
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.10=hcdb4288_3
libffi=3.4.2=h9c3ff4c_2
libflac=1.3.3=h9c3ff4c_1
libgcc-ng=11.2.0=h1d223b6_8
libgfortran-ng=11.2.0=h69a702a_8
libgfortran5=11.2.0=h5c6108e_8
libglib=2.68.4=h174f98d_1
libgomp=11.2.0=h1d223b6_8
libiconv=1.16=h516909a_0
liblapack=3.9.0=11_linux64_mkl
libllvm11=11.1.0=hf817b99_2
libogg=1.3.4=h7f98852_1
libopus=1.3.1=h7f98852_1
libpng=1.6.37=h21135ba_2
libpq=13.3=hd57d9b9_0
librosa=0.8.1=pyhd8ed1ab_0
libsndfile=1.0.31=h9c3ff4c_1
libstdcxx-ng=11.2.0=he4da1e4_8
libtiff=4.3.0=hf544144_1
libtool=2.4.6=h9c3ff4c_1008
libuuid=2.32.1=h7f98852_1000
libuv=1.40.0=h7b6447c_0
libvorbis=1.3.7=h9c3ff4c_0
libwebp-base=1.2.1=h7f98852_0
libxcb=1.13=h7f98852_1003
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.12=h72842e0_0
libxslt=1.1.33=h15afd5d_2
llvmlite=0.37.0=py37h9d7f4d0_0
lz4-c=1.9.3=h9c3ff4c_1
matplotlib-base=3.4.3=py37h1058ff1_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.0=py37h42c9631_2
mkl_random=1.2.2=py37h51133e4_0
mysql-common=8.0.25=ha770c72_2
mysql-libs=8.0.25=hfa10184_2
ncurses=6.2=h58526e2_4
nettle=3.6=he412f7d_0
ninja=1.10.2=hff7bd54_1
nspr=4.30=h9c3ff4c_0
nss=3.69=hb5efdd6_0
numba=0.54.0=py37h2d894fd_0
numpy=1.20.3=py37hf144106_0
numpy-base=1.20.3=py37h74d4b33_0
olefile=0.46=py37_0
omegaconf=2.1.1=py37h89c1867_0
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=1.1.1l=h7f98852_0
packaging=21.0=pyhd8ed1ab_0
pandas=1.3.3=pypi_0
pcre=8.45=h9c3ff4c_0
pillow=8.3.2=py37h0f21c89_0
pip=21.2.4=pyhd8ed1ab_0
pooch=1.5.1=pyhd8ed1ab_0
pthread-stubs=0.4=h36c2ea0_1001
pulseaudio=14.0=hb166930_3
pyannote-core=4.1=pypi_0
pycparser=2.20=pyh9f0ad1d_2
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyside2=5.13.2=py37hfa98aef_4
pysocks=1.7.1=py37h89c1867_3
pysoundfile=0.10.3.post1=pyhd3deb0d_0
python=3.7.10=hb7a2778_101_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-levenshtein=0.12.2=py37h5e8e339_0
python_abi=3.7=2_cp37m
pytorch=1.9.1=py3.7_cpu_0
pytz=2021.1=pypi_0
pyyaml=5.4.1=py37h5e8e339_1
qt=5.12.9=hda022c4_4
randomcolor=0.4.4.6=py37h89c1867_0
readline=8.1=h46c0cb4_0
requests=2.26.0=pyhd8ed1ab_0
resampy=0.2.2=py_0
scikit-learn=0.24.2=py37hf0f1638_1
scipy=1.7.1=py37hf2a6cf1_0
setuptools=58.0.4=py37h89c1867_1
simplejson=3.17.5=pypi_0
six=1.16.0=pyhd3eb1b0_0
sortedcontainers=2.4.0=pypi_0
sqlite=3.36.0=h9cd32fc_1
threadpoolctl=2.2.0=pyh8a188c0_0
tk=8.6.11=h27826a3_1
torchaudio=0.9.1=py37
torchvision=0.2.2=py_3
tornado=6.1=py37h5e8e339_1
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.6=pyhd8ed1ab_0
wheel=0.37.0=pyhd8ed1ab_1
x264=1!161.3030=h7f98852_1
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
zlib=1.2.11=h516909a_1010
zstd=1.5.0=ha95c52a_0
